break down both media_benchmark_tests and media_eval_tests into separate benchmark and eval tests?
